+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.15.0] - 2018-12-13
 ### Added
 - `ParserError::InvalidAttributeValue`.
 - `ElementType::is_filter_primitive`.
@@ -473,7 +475,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/RazrFalcon/svgdom/compare/v0.14.0...HEAD
+[Unreleased]: https://github.com/RazrFalcon/svgdom/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/RazrFalcon/svgdom/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/RazrFalcon/svgdom/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/RazrFalcon/svgdom/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/RazrFalcon/svgdom/compare/v0.11.1...v0.12.0